+++ tpot2/__init__.py
@@ -1,21 +1,14 @@
 
 #TODO: are all the imports in the init files done correctly?
 #TODO clean up import organization
-from ._version import __version__
-from .graphsklearn import *
-from .individual import *
-from .logbook import *
-from .population import *
-from .evolvers import *
-from .objectives import *
-from .evolutionary_algorithms import *
-from .tpot_estimator import *
-from .builtin_modules import *
-from .config import *
-from .representations import *
-from .parent_selectors import *
-from .utils import *
-from .objectives import *
-from .optuna_optimize import *
-from .experimental import *
-from .eval_utils import *
+
+from .graphsklearn import GraphPipeline
+from . import builtin_modules
+from . import config
+from . import individual_representations
+from . import evolvers
+from . import objectives
+from . import selectors
+from . import tpot_estimator
+
+from .tpot_estimator import TPOTClassifier, TPOTRegressor, TPOTEstimator, TPOTEstimatorSteadyState
