--- tpot2/evolutionary_algorithms/__init__.py
@@ -1,2 +1,0 @@
-
-
