--- tpot2/experimental/__init__.py
@@ -1,2 +1,0 @@
-from ..evolvers.steady_state_evolver import *
-from ..tpot_estimator.steady_state_estimator import *
