+++ tpot2/individual_representations/__init__.py
@@ -1,2 +1,3 @@
+from .individual import BaseIndividual
 from .subset_selector import SubsetSelector
 from .graph_pipeline_individual import *
