+++ tpot2/individual_representations/graph_pipeline_individual/templates.py
@@ -2,10 +2,10 @@
 import numpy as np
 import tpot2
 import networkx as nx
-from tpot2.representations.graph_pipeline_individual import GraphIndividual
+from tpot2.individual_representations.graph_pipeline_individual import GraphIndividual
 import random
 
-from tpot2.representations.graph_pipeline_individual.individual import create_node
+from tpot2.individual_representations.graph_pipeline_individual.individual import create_node
 
 
 def estimator_graph_individual_generator(
