+++ tpot2/individual_representations/individual.py
