+++ tpot2/objectives/number_of_leaves_scorer.py
@@ -1,0 +1,2 @@
+def number_of_leaves_scorer(est,X,y):
+    return len([v for v, d in est.graph.out_degree() if d == 0])
