+++ tpot2/objectives/number_of_nodes_objective.py
@@ -1,0 +1,2 @@
+def number_of_nodes_objective(graph_pipeline):
+    return graph_pipeline.graph.number_of_nodes()
