+++ tpot2/tpot_estimator/steady_state_estimator.py
@@ -1,19 +1,16 @@
 from sklearn.base import BaseEstimator
 from sklearn.utils.metaestimators import available_if
 import numpy as np
-import typing
 import sklearn.metrics
-import tpot2.objectives.estimator_objective_functions
-from functools import partial
 import tpot2.config
-from sklearn.utils.validation import check_X_y, check_array, check_is_fitted
-from tpot2.parent_selectors import survival_select_NSGA2, TournamentSelection_Dominated
+from sklearn.utils.validation import  check_is_fitted
+from tpot2.selectors import survival_select_NSGA2, tournament_selection_dominated
 from sklearn.preprocessing import LabelEncoder 
 from sklearn.utils.multiclass import unique_labels 
 import pandas as pd
 from sklearn.model_selection import train_test_split
 import tpot2
-import distributed
+
 from dask.distributed import Client
 from dask.distributed import LocalCluster
 import math
@@ -33,8 +30,8 @@ class TPOTEstimatorSteadyState(BaseEstimator):
                         scorers_weights = [],
                         classification = False,
                         cv = 5,
-                        other_objective_functions=[tpot2.objectives.estimator_objective_functions.average_path_length_objective], #tpot2.objectives.estimator_objective_functions.number_of_nodes_objective],
-                        other_objective_functions_weights = [-1],
+                        other_objective_functions=[], #tpot2.objectives.estimator_objective_functions.number_of_nodes_objective],
+                        other_objective_functions_weights = [],
                         objective_function_names = None,
                         bigger_is_better = True,
                         max_size = np.inf, 
@@ -72,7 +69,7 @@ class TPOTEstimatorSteadyState(BaseEstimator):
                         mutate_then_crossover_probability=.05,
                         crossover_then_mutate_probability=.05,
                         survival_selector = survival_select_NSGA2,
-                        parent_selector = TournamentSelection_Dominated,
+                        parent_selector = tournament_selection_dominated,
                         budget_range = None,
                         budget_scaling = .5,
                         individuals_until_end_budget = 1,  
@@ -83,7 +80,7 @@ class TPOTEstimatorSteadyState(BaseEstimator):
 
                         verbose = 0,
                         periodic_checkpoint_folder = None, 
-                        callback: tpot2.CallBackInterface = None,
+                        callback = None,
                         processes = True,
 
                         scatter = True,
